fix(R.10): close QA-4 changed-file findings
