+++ crates/atm/src/commands/ack.rs
@@ -59,15 +59,17 @@ impl AckCommand {
 
 #[cfg(test)]
 mod tests {
-    use std::path::PathBuf;
+    use tempfile::TempDir;
 
     use super::AckCommand;
 
     const VALID_MESSAGE_ID: &str = "550e8400-e29b-41d4-a716-446655440000";
 
-    fn test_paths() -> (PathBuf, PathBuf) {
-        let root = std::env::temp_dir().join("atm-ack-command-tests");
-        (root.join("home"), root.join("cwd"))
+    fn test_paths() -> (TempDir, std::path::PathBuf, std::path::PathBuf) {
+        let tempdir = TempDir::new().expect("tempdir");
+        let home_dir = tempdir.path().join("home");
+        let current_dir = tempdir.path().join("cwd");
+        (tempdir, home_dir, current_dir)
     }
 
     #[test]
@@ -80,7 +82,7 @@ mod tests {
             json: false,
         };
 
-        let (home_dir, current_dir) = test_paths();
+        let (_tempdir, home_dir, current_dir) = test_paths();
         let error = command
             .build_request(home_dir, current_dir)
             .expect_err("empty message id");
@@ -98,7 +100,7 @@ mod tests {
             json: false,
         };
 
-        let (home_dir, current_dir) = test_paths();
+        let (_tempdir, home_dir, current_dir) = test_paths();
         let error = command
             .build_request(home_dir, current_dir)
             .expect_err("whitespace message id");
@@ -116,7 +118,7 @@ mod tests {
             json: true,
         };
 
-        let (home_dir, current_dir) = test_paths();
+        let (_tempdir, home_dir, current_dir) = test_paths();
         let request = command
             .build_request(home_dir, current_dir)
             .expect("request");
